Fix POWER compilation wrt __u64 printf/scanf
On POWER, __u64 is defined in the Linux kernel headers as
"long unsinged", while on x86_64 it is defined as "long long unsigned".
As a result, we have no way to uniformly printf/scanf __u64 values
without type casting.

The closest userspace type equivalent of __u64 is uint64_t
(as defined by ISO C99: 7.18 Integer types <stdint.h>),
and this is what we should cast it to. Unfortunately, uint64_t
can also be either "unsigned long" or "unsigned long long",
depending on the platform, defines etc.

Therefore, it seems the best way to pass a __u64 value to printf/scanf
is to cast it to uint64_t and use PRIu64 / SCNu64 macros (as defined
by ISO C99: 7.8 Format conversion of integer types <inttypes.h>).

In most cases we should keep using __u64 type as this is what kernel
uses and we are working with the kernel. In some cases, though, we
need to switch to using uint64_t so we don't need to cast the values
passed on to printf/scanf.

This fixes a bunch of errors like:

> ploop.c:1721:2: error: format ‘%llu’ expects argument of type ‘long long unsigned int’, but argument 3 has type ‘long unsigned int’ [-Werror=format=]
>   ploop_log(0, "Setting maximum delta size to %llu sec", size);
>   ^

and

> fsutils.c:246:5: error: format ‘%llu’ expects argument of type ‘long long unsigned int *’, but argument 3 has type ‘__u64 *’ [-Werror=format=]
>     sscanf(buf, "Free blocks: %llu", &data->block_free) == 1)
>     ^

+++ lib/balloon_util.c
@@ -271,16 +271,16 @@ static int fiemap_extent_process(__u32 clu, __u32 len, __u32 *rmap, __u32 rlen,
 							delta->blocksize, delta->version);
 			if (ridx >= rlen) {
 				ploop_err(0,
-					"Image corrupted: L2[%u] == %u (max=%llu)",
+					"Image corrupted: L2[%u] == %u (max=%" PRIu64 ")",
 					clu + j - l2_slot, delta->l2[j],
-					(rlen - 1) * B2S(cluster));
+					(uint64_t)((rlen - 1) * B2S(cluster)));
 				return(SYSEXIT_PLOOPFMT);
 			}
 			if (ridx < delta->l1_size) {
 				ploop_err(0,
-					"Image corrupted: L2[%u] == %u (min=%llu)",
+					"Image corrupted: L2[%u] == %u (min=%" PRIu64 ")",
 					clu + j - l2_slot, delta->l2[j],
-					delta->l1_size * B2S(cluster));
+					(uint64_t)(delta->l1_size * B2S(cluster)));
 				return(SYSEXIT_PLOOPFMT);
 			}
 
@@ -536,16 +536,16 @@ static int range_build_rmap(__u32 iblk_start, __u32 iblk_end,
 				delta->blocksize, delta->version);
 		if (ridx >= rlen) {
 			ploop_err(0,
-				"Image corrupted: L2[%u] == %u (max=%llu) (2)",
+				"Image corrupted: L2[%u] == %u (max=%" PRIu64 ") (2)",
 				clu, delta->l2[l2_slot],
-				(rlen - 1) * B2S(cluster));
+				(uint64_t)((rlen - 1) * B2S(cluster)));
 			return SYSEXIT_PLOOPFMT;
 		}
 		if (ridx && ridx < delta->l1_size) {
 			ploop_err(0,
-				"Image corrupted: L2[%u] == %u (min=%llu) (2)",
+				"Image corrupted: L2[%u] == %u (min=%" PRIu64 ") (2)",
 				clu, delta->l2[l2_slot],
-				delta->l1_size * B2S(cluster));
+				(uint64_t)(delta->l1_size * B2S(cluster)));
 			return SYSEXIT_PLOOPFMT;
 		}
